Add `arch_driver` Component, refactor
