+++ Sming/Arch/Esp32/Components/driver/component.mk
@@ -1,9 +1,11 @@
-COMPONENT_DEPENDS			:= esp32
+COMPONENT_DEPENDS := \
+	arch_driver \
+	esp32
 
 COMPONENT_SRCDIRS			:= .
 COMPONENT_INCDIRS			:= include
 
-COMPONENT_DOXYGEN_INPUT		        := include/driver
+COMPONENT_DOXYGEN_INPUT		:= include/driver
 
 # hw_timer
 COMPONENT_VARS				+= USE_US_TIMER
