+++ Sming/Arch/Esp32/Components/driver/include/driver/os_timer.h
@@ -6,15 +6,16 @@
  *
  * os_timer.h
  *
- * @author: 13 August 2018 - mikee47 <mike@sillyhouse.net>
+ * This implementation mimics the behaviour of the ESP8266 Non-OS SDK timers,
+ * using Timer2 as the reference (which is _not_ in microseconds!)
  *
- * An alternative method for setting software timers based on the tick count.
- *
+ * The ESP32 IDF contains more sophisticated timer implementations, but also
+ * this same API which it refers to as the 'legacy' timer API.
  */
 
 #pragma once
 
-#include "esp_systemapi.h"
+#include <esp32/rom/ets_sys.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -26,6 +27,14 @@ extern "C" {
  * @{
  */
 
+typedef ETSTimerFunc os_timer_func_t;
+typedef ETSTimer os_timer_t;
+
+#define os_timer_arm(ptimer, ms, repeat_flag) ets_timer_arm(ptimer, ms, repeat_flag)
+#define os_timer_arm_us(ptimer, us, repeat_flag) ets_timer_arm_us(ptimer, us, repeat_flag)
+#define os_timer_disarm(ptimer) ets_timer_disarm(ptimer)
+#define os_timer_setfn(ptimer, pfunction, parg) ets_timer_setfn(ptimer, pfunction, parg)
+
 /**
  * @brief Set a software timer using the Timer2 tick value
  * @param ptimer Timer structure
@@ -35,7 +44,7 @@ extern "C" {
  * This function has been added to Sming for more efficient and flexible use of
  * software timers. It can be used alongside the SDK `os_timer_arm_new()` function.
  */
-void IRAM_ATTR os_timer_arm_ticks(os_timer_t* ptimer, uint32_t ticks, bool repeat_flag);
+void os_timer_arm_ticks(os_timer_t* ptimer, uint32_t ticks, bool repeat_flag);
 
 /** @} */
