+++ Sming/Arch/Esp32/Components/driver/os_timer.cpp
@@ -1,0 +1,7 @@
+#include <driver/os_timer.h>
+#include <esp_attr.h>
+
+void IRAM_ATTR os_timer_arm_ticks(os_timer_t* ptimer, uint32_t ticks, bool repeat_flag)
+{
+	ets_timer_arm(ptimer, ticks / 5000, repeat_flag);
+}
