+++ Sming/Arch/Esp32/Components/esp32/src/include/esp_systemapi.h
@@ -25,7 +25,6 @@
 #include "esp_sleep.h"
 #include "esp_clk.h"
 #include "esp_tasks.h"
-#include "esp_timer_legacy.h"
 #include <heap.h>
 #include <esp_system.h>
 #include <espinc/pin_mux_register.h>
