+++ Sming/Arch/Esp32/Core/adc.cpp
@@ -1,5 +1,6 @@
 #include <driver/adc.h>
 #include <esp_adc_cal.h>
+#include <debug_progmem.h>
 
 #define DEFAULT_VREF 1100 //Use adc2_vref_to_gpio() to obtain a better estimate
