+++ Sming/Arch/Esp32/Core/peripheral.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#if 0
+
 #ifdef __cplusplus
 extern "C" {
 #endif
@@ -76,10 +78,10 @@ extern "C" {
 #define GPC15 ESP8266_REG(0x364) //GPIO_PIN15
 
 //GPIO (0-15) PIN Control Bits
-#define GPCWE 10 //WAKEUP_ENABLE (can be 1 only when INT_TYPE is high or low)
-#define GPCI 7   //INT_TYPE (3bits) 0:disable,1:rising,2:falling,3:change,4:low,5:high
-#define GPCD 2   //DRIVER 0:normal,1:open drain
-#define GPCS 0   //SOURCE 0:GPIO_DATA,1:SigmaDelta
+#define GPCWE 10				 //WAKEUP_ENABLE (can be 1 only when INT_TYPE is high or low)
+#define GPCI 7					 //INT_TYPE (3bits) 0:disable,1:rising,2:falling,3:change,4:low,5:high
+#define GPCD 2					 //DRIVER 0:normal,1:open drain
+#define GPCS 0					 //SOURCE 0:GPIO_DATA,1:SigmaDelta
 
 #define GPMUX ESP8266_REG(0x800)
 //GPIO (0-15) PIN Function Registers
@@ -147,21 +149,21 @@ extern const uint8_t esp8266_gpioToFn[16];
 #define T1I ESP8266_REG(0x60C) //Interrupt Status Register (1bit) write to clear
 //edge interrupt enable register
 #define TEIE ESP8266_DREG(0x04)
-#define TEIE1 0x02 //bit for timer 1
+#define TEIE1 0x02				 //bit for timer 1
 
 //Timer 2 Registers (32bit CountUp Timer)
-#define T2L ESP8266_REG(0x620) //Load Value (Starting Value of Counter)
-#define T2V ESP8266_REG(0x624) //(RO) Current Value
-#define T2C ESP8266_REG(0x628) //Control Register
-#define T2I ESP8266_REG(0x62C) //Interrupt Status Register (1bit) write to clear
-#define T2A ESP8266_REG(0x630) //Alarm Value
+#define T2L ESP8266_REG(0x620)   //Load Value (Starting Value of Counter)
+#define T2V ESP8266_REG(0x624)   //(RO) Current Value
+#define T2C ESP8266_REG(0x628)   //Control Register
+#define T2I ESP8266_REG(0x62C)   //Interrupt Status Register (1bit) write to clear
+#define T2A ESP8266_REG(0x630)   //Alarm Value
 
 //Timer Control Bits
-#define TCIS 8 //Interrupt Status
-#define TCTE 7 //Timer Enable
-#define TCAR 6 //AutoReload (restart timer when condition is reached)
-#define TCPD 2 //Prescale Devider (2bit) 0:1(12.5ns/tick), 1:16(0.2us/tick), 2/3:256(3.2us/tick)
-#define TCIT 0 //Interrupt Type 0:edge, 1:level
+#define TCIS 8					 //Interrupt Status
+#define TCTE 7					 //Timer Enable
+#define TCAR 6					 //AutoReload (restart timer when condition is reached)
+#define TCPD 2					 //Prescale Devider (2bit) 0:1(12.5ns/tick), 1:16(0.2us/tick), 2/3:256(3.2us/tick)
+#define TCIT 0					 //Interrupt Type 0:edge, 1:level
 
 //RTC Registers
 #define RTCSV ESP8266_REG(0x704) //RTC SLEEP COUNTER Target Value
@@ -188,38 +190,38 @@ extern const uint8_t esp8266_gpioToFn[16];
 #define UIS1 2
 
 //UART 0 Registers
-#define U0F ESP8266_REG(0x000)  //UART FIFO
-#define U0IR ESP8266_REG(0x004) //INT_RAW
-#define U0IS ESP8266_REG(0x008) //INT_STATUS
-#define U0IE ESP8266_REG(0x00c) //INT_ENABLE
-#define U0IC ESP8266_REG(0x010) //INT_CLEAR
-#define U0D ESP8266_REG(0x014)  //CLKDIV
-#define U0A ESP8266_REG(0x018)  //AUTOBAUD
-#define U0S ESP8266_REG(0x01C)  //STATUS
-#define U0C0 ESP8266_REG(0x020) //CONF0
-#define U0C1 ESP8266_REG(0x024) //CONF1
-#define U0LP ESP8266_REG(0x028) //LOW_PULSE
-#define U0HP ESP8266_REG(0x02C) //HIGH_PULSE
-#define U0PN ESP8266_REG(0x030) //PULSE_NUM
-#define U0DT ESP8266_REG(0x078) //DATE
-#define U0ID ESP8266_REG(0x07C) //ID
+#define U0F ESP8266_REG(0x000)						   //UART FIFO
+#define U0IR ESP8266_REG(0x004)						   //INT_RAW
+#define U0IS ESP8266_REG(0x008)						   //INT_STATUS
+#define U0IE ESP8266_REG(0x00c)						   //INT_ENABLE
+#define U0IC ESP8266_REG(0x010)						   //INT_CLEAR
+#define U0D ESP8266_REG(0x014)						   //CLKDIV
+#define U0A ESP8266_REG(0x018)						   //AUTOBAUD
+#define U0S ESP8266_REG(0x01C)						   //STATUS
+#define U0C0 ESP8266_REG(0x020)						   //CONF0
+#define U0C1 ESP8266_REG(0x024)						   //CONF1
+#define U0LP ESP8266_REG(0x028)						   //LOW_PULSE
+#define U0HP ESP8266_REG(0x02C)						   //HIGH_PULSE
+#define U0PN ESP8266_REG(0x030)						   //PULSE_NUM
+#define U0DT ESP8266_REG(0x078)						   //DATE
+#define U0ID ESP8266_REG(0x07C)						   //ID
 
 //UART 1 Registers
-#define U1F ESP8266_REG(0xF00)  //UART FIFO
-#define U1IR ESP8266_REG(0xF04) //INT_RAW
-#define U1IS ESP8266_REG(0xF08) //INT_STATUS
-#define U1IE ESP8266_REG(0xF0c) //INT_ENABLE
-#define U1IC ESP8266_REG(0xF10) //INT_CLEAR
-#define U1D ESP8266_REG(0xF14)  //CLKDIV
-#define U1A ESP8266_REG(0xF18)  //AUTOBAUD
-#define U1S ESP8266_REG(0xF1C)  //STATUS
-#define U1C0 ESP8266_REG(0xF20) //CONF0
-#define U1C1 ESP8266_REG(0xF24) //CONF1
-#define U1LP ESP8266_REG(0xF28) //LOW_PULSE
-#define U1HP ESP8266_REG(0xF2C) //HIGH_PULSE
-#define U1PN ESP8266_REG(0xF30) //PULSE_NUM
-#define U1DT ESP8266_REG(0xF78) //DATE
-#define U1ID ESP8266_REG(0xF7C) //ID
+#define U1F ESP8266_REG(0xF00)						   //UART FIFO
+#define U1IR ESP8266_REG(0xF04)						   //INT_RAW
+#define U1IS ESP8266_REG(0xF08)						   //INT_STATUS
+#define U1IE ESP8266_REG(0xF0c)						   //INT_ENABLE
+#define U1IC ESP8266_REG(0xF10)						   //INT_CLEAR
+#define U1D ESP8266_REG(0xF14)						   //CLKDIV
+#define U1A ESP8266_REG(0xF18)						   //AUTOBAUD
+#define U1S ESP8266_REG(0xF1C)						   //STATUS
+#define U1C0 ESP8266_REG(0xF20)						   //CONF0
+#define U1C1 ESP8266_REG(0xF24)						   //CONF1
+#define U1LP ESP8266_REG(0xF28)						   //LOW_PULSE
+#define U1HP ESP8266_REG(0xF2C)						   //HIGH_PULSE
+#define U1PN ESP8266_REG(0xF30)						   //PULSE_NUM
+#define U1DT ESP8266_REG(0xF78)						   //DATE
+#define U1ID ESP8266_REG(0xF7C)						   //ID
 
 //UART(uart) Registers
 #define USF(u) ESP8266_REG(0x000 + (0xF00 * (u & 1)))  //UART FIFO
@@ -239,52 +241,52 @@ extern const uint8_t esp8266_gpioToFn[16];
 #define USID(u) ESP8266_REG(0x07C + (0xF00 * (u & 1))) //ID
 
 //UART INT Registers Bits
-#define UITO 8  //RX FIFO TimeOut
-#define UIBD 7  //Break Detected
-#define UICTS 6 //CTS Changed
-#define UIDSR 5 //DSR Change
-#define UIOF 4  //RX FIFO OverFlow
-#define UIFR 3  //Frame Error
-#define UIPE 2  //Parity Error
-#define UIFE 1  //TX FIFO Empty
-#define UIFF 0  //RX FIFO Full
+#define UITO 8										   //RX FIFO TimeOut
+#define UIBD 7										   //Break Detected
+#define UICTS 6										   //CTS Changed
+#define UIDSR 5										   //DSR Change
+#define UIOF 4										   //RX FIFO OverFlow
+#define UIFR 3										   //Frame Error
+#define UIPE 2										   //Parity Error
+#define UIFE 1										   //TX FIFO Empty
+#define UIFF 0										   //RX FIFO Full
 
 //UART STATUS Registers Bits
-#define USTX 31  //TX PIN Level
-#define USRTS 30 //RTS PIN Level
-#define USDTR 39 //DTR PIN Level
-#define USTXC 16 //TX FIFO COUNT (8bit)
-#define USRXD 15 //RX PIN Level
-#define USCTS 14 //CTS PIN Level
-#define USDSR 13 //DSR PIN Level
-#define USRXC 0  //RX FIFO COUNT (8bit)
+#define USTX 31										   //TX PIN Level
+#define USRTS 30									   //RTS PIN Level
+#define USDTR 39									   //DTR PIN Level
+#define USTXC 16									   //TX FIFO COUNT (8bit)
+#define USRXD 15									   //RX PIN Level
+#define USCTS 14									   //CTS PIN Level
+#define USDSR 13									   //DSR PIN Level
+#define USRXC 0										   //RX FIFO COUNT (8bit)
 
 //UART CONF0 Registers Bits
-#define UCDTRI 24  //Invert DTR
-#define UCRTSI 23  //Invert RTS
-#define UCTXI 22   //Invert TX
-#define UCDSRI 21  //Invert DSR
-#define UCCTSI 20  //Invert CTS
-#define UCRXI 19   //Invert RX
-#define UCTXRST 18 //Reset TX FIFO
-#define UCRXRST 17 //Reset RX FIFO
-#define UCTXHFE 15 //TX Harware Flow Enable
-#define UCLBE 14   //LoopBack Enable
-#define UCBRK 8	//Send Break on the TX line
-#define UCSWDTR 7  //Set this bit to assert DTR
-#define UCSWRTS 6  //Set this bit to assert RTS
-#define UCSBN 4	//StopBits Count (2bit) 0:disable, 1:1bit, 2:1.5bit, 3:2bit
-#define UCBN 2	 //DataBits Count (2bin) 0:5bit, 1:6bit, 2:7bit, 3:8bit
-#define UCPAE 1	//Parity Enable
-#define UCPA 0	 //Parity 0:even, 1:odd
+#define UCDTRI 24									   //Invert DTR
+#define UCRTSI 23									   //Invert RTS
+#define UCTXI 22									   //Invert TX
+#define UCDSRI 21									   //Invert DSR
+#define UCCTSI 20									   //Invert CTS
+#define UCRXI 19									   //Invert RX
+#define UCTXRST 18									   //Reset TX FIFO
+#define UCRXRST 17									   //Reset RX FIFO
+#define UCTXHFE 15									   //TX Harware Flow Enable
+#define UCLBE 14									   //LoopBack Enable
+#define UCBRK 8										   //Send Break on the TX line
+#define UCSWDTR 7									   //Set this bit to assert DTR
+#define UCSWRTS 6									   //Set this bit to assert RTS
+#define UCSBN 4										   //StopBits Count (2bit) 0:disable, 1:1bit, 2:1.5bit, 3:2bit
+#define UCBN 2										   //DataBits Count (2bin) 0:5bit, 1:6bit, 2:7bit, 3:8bit
+#define UCPAE 1										   //Parity Enable
+#define UCPA 0										   //Parity 0:even, 1:odd
 
 //UART CONF1 Registers Bits
-#define UCTOE 31   //RX TimeOut Enable
-#define UCTOT 24   //RX TimeOut Treshold (7bit)
-#define UCRXHFE 23 //RX Harware Flow Enable
-#define UCRXHFT 16 //RX Harware Flow Treshold (7bit)
-#define UCFET 8	//TX FIFO Empty Treshold (7bit)
-#define UCFFT 0	//RX FIFO Full Treshold (7bit)
+#define UCTOE 31									   //RX TimeOut Enable
+#define UCTOT 24									   //RX TimeOut Treshold (7bit)
+#define UCRXHFE 23									   //RX Harware Flow Enable
+#define UCRXHFT 16									   //RX Harware Flow Treshold (7bit)
+#define UCFET 8										   //TX FIFO Empty Treshold (7bit)
+#define UCFFT 0										   //RX FIFO Full Treshold (7bit)
 
 //SPI_READY
 #define SPIRDY ESP8266_DREG(0x0C)
@@ -367,85 +369,85 @@ extern const uint8_t esp8266_gpioToFn[16];
 
 //SPI0, SPI1 & I2S Interupt Register
 #define SPIIR ESP8266_DREG(0x20)
-#define SPII0 4 //SPI0 Interrupt
-#define SPII1 7 //SPI1 Interrupt
-#define SPII2 9 //I2S Interrupt
+#define SPII0 4						//SPI0 Interrupt
+#define SPII1 7						//SPI1 Interrupt
+#define SPII2 9						//I2S Interrupt
 
 //SPI CMD
-#define SPICMDREAD (1 << 31) //SPI_FLASH_READ
-#define SPICMDWREN (1 << 30) //SPI_FLASH_WREN
-#define SPICMDWRDI (1 << 29) //SPI_FLASH_WRDI
-#define SPICMDRDID (1 << 28) //SPI_FLASH_RDID
-#define SPICMDRDSR (1 << 27) //SPI_FLASH_RDSR
-#define SPICMDWRSR (1 << 26) //SPI_FLASH_WRSR
-#define SPICMDPP (1 << 25)   //SPI_FLASH_PP
-#define SPICMDSE (1 << 24)   //SPI_FLASH_SE
-#define SPICMDBE (1 << 23)   //SPI_FLASH_BE
-#define SPICMDCE (1 << 22)   //SPI_FLASH_CE
-#define SPICMDDP (1 << 21)   //SPI_FLASH_DP
-#define SPICMDRES (1 << 20)  //SPI_FLASH_RES
-#define SPICMDHPM (1 << 19)  //SPI_FLASH_HPM
-#define SPICMDUSR (1 << 18)  //SPI_FLASH_USR
-#define SPIBUSY (1 << 18)	//SPI_USR
+#define SPICMDREAD (1 << 31)		//SPI_FLASH_READ
+#define SPICMDWREN (1 << 30)		//SPI_FLASH_WREN
+#define SPICMDWRDI (1 << 29)		//SPI_FLASH_WRDI
+#define SPICMDRDID (1 << 28)		//SPI_FLASH_RDID
+#define SPICMDRDSR (1 << 27)		//SPI_FLASH_RDSR
+#define SPICMDWRSR (1 << 26)		//SPI_FLASH_WRSR
+#define SPICMDPP (1 << 25)			//SPI_FLASH_PP
+#define SPICMDSE (1 << 24)			//SPI_FLASH_SE
+#define SPICMDBE (1 << 23)			//SPI_FLASH_BE
+#define SPICMDCE (1 << 22)			//SPI_FLASH_CE
+#define SPICMDDP (1 << 21)			//SPI_FLASH_DP
+#define SPICMDRES (1 << 20)			//SPI_FLASH_RES
+#define SPICMDHPM (1 << 19)			//SPI_FLASH_HPM
+#define SPICMDUSR (1 << 18)			//SPI_FLASH_USR
+#define SPIBUSY (1 << 18)			//SPI_USR
 
 //SPI CTRL (SPIxC)
-#define SPICWBO (1 << 26)		//SPI_WR_BIT_ODER
-#define SPICRBO (1 << 25)		//SPI_RD_BIT_ODER
-#define SPICQIO (1 << 24)		//SPI_QIO_MODE
-#define SPICDIO (1 << 23)		//SPI_DIO_MODE
-#define SPIC2BSE (1 << 22)		//SPI_TWO_BYTE_STATUS_EN
-#define SPICWPR (1 << 21)		//SPI_WP_REG
-#define SPICQOUT (1 << 20)		//SPI_QOUT_MODE
-#define SPICSHARE (1 << 19)		//SPI_SHARE_BUS
-#define SPICHOLD (1 << 18)		//SPI_HOLD_MODE
-#define SPICAHB (1 << 17)		//SPI_ENABLE_AHB
-#define SPICSSTAAI (1 << 16)	//SPI_SST_AAI
-#define SPICRESANDRES (1 << 15) //SPI_RESANDRES
-#define SPICDOUT (1 << 14)		//SPI_DOUT_MODE
-#define SPICFASTRD (1 << 13)	//SPI_FASTRD_MODE
+#define SPICWBO (1 << 26)			//SPI_WR_BIT_ODER
+#define SPICRBO (1 << 25)			//SPI_RD_BIT_ODER
+#define SPICQIO (1 << 24)			//SPI_QIO_MODE
+#define SPICDIO (1 << 23)			//SPI_DIO_MODE
+#define SPIC2BSE (1 << 22)			//SPI_TWO_BYTE_STATUS_EN
+#define SPICWPR (1 << 21)			//SPI_WP_REG
+#define SPICQOUT (1 << 20)			//SPI_QOUT_MODE
+#define SPICSHARE (1 << 19)			//SPI_SHARE_BUS
+#define SPICHOLD (1 << 18)			//SPI_HOLD_MODE
+#define SPICAHB (1 << 17)			//SPI_ENABLE_AHB
+#define SPICSSTAAI (1 << 16)		//SPI_SST_AAI
+#define SPICRESANDRES (1 << 15)		//SPI_RESANDRES
+#define SPICDOUT (1 << 14)			//SPI_DOUT_MODE
+#define SPICFASTRD (1 << 13)		//SPI_FASTRD_MODE
 
 //SPI CTRL1 (SPIxC1)
-#define SPIC1TCSH 0xF   //SPI_T_CSH
-#define SPIC1TCSH_S 28  //SPI_T_CSH_S
-#define SPIC1TRES 0xFFF //SPI_T_RES
-#define SPIC1TRES_S 16  //SPI_T_RES_S
-#define SPIC1BTL 0xFFFF //SPI_BUS_TIMER_LIMIT
-#define SPIC1BTL_S 0	//SPI_BUS_TIMER_LIMIT_S
+#define SPIC1TCSH 0xF				//SPI_T_CSH
+#define SPIC1TCSH_S 28				//SPI_T_CSH_S
+#define SPIC1TRES 0xFFF				//SPI_T_RES
+#define SPIC1TRES_S 16				//SPI_T_RES_S
+#define SPIC1BTL 0xFFFF				//SPI_BUS_TIMER_LIMIT
+#define SPIC1BTL_S 0				//SPI_BUS_TIMER_LIMIT_S
 
 //SPI Status (SPIxRS)
-#define SPIRSEXT 0xFF	  //SPI_STATUS_EXT
-#define SPIRSEXT_S 24	  //SPI_STATUS_EXT_S
-#define SPIRSWB 0xFF	   //SPI_WB_MODE
-#define SPIRSWB_S 16	   //SPI_WB_MODE_S
-#define SPIRSSP (1 << 7)   //SPI_FLASH_STATUS_PRO_FLAG
-#define SPIRSTBP (1 << 5)  //SPI_FLASH_TOP_BOT_PRO_FLAG
-#define SPIRSBP2 (1 << 4)  //SPI_FLASH_BP2
-#define SPIRSBP1 (1 << 3)  //SPI_FLASH_BP1
-#define SPIRSBP0 (1 << 2)  //SPI_FLASH_BP0
-#define SPIRSWRE (1 << 1)  //SPI_FLASH_WRENABLE_FLAG
-#define SPIRSBUSY (1 << 0) //SPI_FLASH_BUSY_FLAG
+#define SPIRSEXT 0xFF				//SPI_STATUS_EXT
+#define SPIRSEXT_S 24				//SPI_STATUS_EXT_S
+#define SPIRSWB 0xFF				//SPI_WB_MODE
+#define SPIRSWB_S 16				//SPI_WB_MODE_S
+#define SPIRSSP (1 << 7)			//SPI_FLASH_STATUS_PRO_FLAG
+#define SPIRSTBP (1 << 5)			//SPI_FLASH_TOP_BOT_PRO_FLAG
+#define SPIRSBP2 (1 << 4)			//SPI_FLASH_BP2
+#define SPIRSBP1 (1 << 3)			//SPI_FLASH_BP1
+#define SPIRSBP0 (1 << 2)			//SPI_FLASH_BP0
+#define SPIRSWRE (1 << 1)			//SPI_FLASH_WRENABLE_FLAG
+#define SPIRSBUSY (1 << 0)			//SPI_FLASH_BUSY_FLAG
 
 //SPI CTRL2 (SPIxC2)
-#define SPIC2CSDN 0xF	//SPI_CS_DELAY_NUM
-#define SPIC2CSDN_S 28   //SPI_CS_DELAY_NUM_S
-#define SPIC2CSDM 0x3	//SPI_CS_DELAY_MODE
-#define SPIC2CSDM_S 26   //SPI_CS_DELAY_MODE_S
-#define SPIC2MOSIDN 0x7  //SPI_MOSI_DELAY_NUM
-#define SPIC2MOSIDN_S 23 //SPI_MOSI_DELAY_NUM_S
-#define SPIC2MOSIDM 0x3  //SPI_MOSI_DELAY_MODE
-#define SPIC2MOSIDM_S 21 //SPI_MOSI_DELAY_MODE_S
-#define SPIC2MISODN 0x7  //SPI_MISO_DELAY_NUM
-#define SPIC2MISODN_S 18 //SPI_MISO_DELAY_NUM_S
-#define SPIC2MISODM 0x3  //SPI_MISO_DELAY_MODE
-#define SPIC2MISODM_S 16 //SPI_MISO_DELAY_MODE_S
-#define SPIC2CKOHM 0xF   //SPI_CK_OUT_HIGH_MODE
-#define SPIC2CKOHM_S 12  //SPI_CK_OUT_HIGH_MODE_S
-#define SPIC2CKOLM 0xF   //SPI_CK_OUT_LOW_MODE
-#define SPIC2CKOLM_S 8   //SPI_CK_OUT_LOW_MODE_S
-#define SPIC2HT 0xF		 //SPI_HOLD_TIME
-#define SPIC2HT_S 4		 //SPI_HOLD_TIME_S
-#define SPIC2ST 0xF		 //SPI_SETUP_TIME
-#define SPIC2ST_S 0		 //SPI_SETUP_TIME_S
+#define SPIC2CSDN 0xF				//SPI_CS_DELAY_NUM
+#define SPIC2CSDN_S 28				//SPI_CS_DELAY_NUM_S
+#define SPIC2CSDM 0x3				//SPI_CS_DELAY_MODE
+#define SPIC2CSDM_S 26				//SPI_CS_DELAY_MODE_S
+#define SPIC2MOSIDN 0x7				//SPI_MOSI_DELAY_NUM
+#define SPIC2MOSIDN_S 23			//SPI_MOSI_DELAY_NUM_S
+#define SPIC2MOSIDM 0x3				//SPI_MOSI_DELAY_MODE
+#define SPIC2MOSIDM_S 21			//SPI_MOSI_DELAY_MODE_S
+#define SPIC2MISODN 0x7				//SPI_MISO_DELAY_NUM
+#define SPIC2MISODN_S 18			//SPI_MISO_DELAY_NUM_S
+#define SPIC2MISODM 0x3				//SPI_MISO_DELAY_MODE
+#define SPIC2MISODM_S 16			//SPI_MISO_DELAY_MODE_S
+#define SPIC2CKOHM 0xF				//SPI_CK_OUT_HIGH_MODE
+#define SPIC2CKOHM_S 12				//SPI_CK_OUT_HIGH_MODE_S
+#define SPIC2CKOLM 0xF				//SPI_CK_OUT_LOW_MODE
+#define SPIC2CKOLM_S 8				//SPI_CK_OUT_LOW_MODE_S
+#define SPIC2HT 0xF					//SPI_HOLD_TIME
+#define SPIC2HT_S 4					//SPI_HOLD_TIME_S
+#define SPIC2ST 0xF					//SPI_SETUP_TIME
+#define SPIC2ST_S 0					//SPI_SETUP_TIME_S
 
 //SPI CLK (SPIxCLK)
 #define SPICLK_EQU_SYSCLK (1 << 31) //SPI_CLK_EQU_SYSCLK
@@ -459,44 +461,44 @@ extern const uint8_t esp8266_gpioToFn[16];
 #define SPICLKCL_S 0				//SPI_CLKCNT_L_S
 
 //SPI Phases (SPIxU)
-#define SPIUCOMMAND (1 << 31)   //COMMAND pahse, SPI_USR_COMMAND
-#define SPIUADDR (1 << 30)		//ADDRESS phase, SPI_FLASH_USR_ADDR
-#define SPIUDUMMY (1 << 29)		//DUMMY phase, SPI_FLASH_USR_DUMMY
-#define SPIUMISO (1 << 28)		//MISO phase, SPI_FLASH_USR_DIN
-#define SPIUMOSI (1 << 27)		//MOSI phase, SPI_FLASH_DOUT
-#define SPIUDUMMYIDLE (1 << 26) //SPI_USR_DUMMY_IDLE
-#define SPIUMOSIH (1 << 25)		//MOSI phase uses W8-W15, SPI_USR_DOUT_HIGHPART
-#define SPIUMISOH (1 << 24)		//MISO pahse uses W8-W15, SPI_USR_DIN_HIGHPART
-#define SPIUPREPHOLD (1 << 23)  //SPI_USR_PREP_HOLD
-#define SPIUCMDHOLD (1 << 22)   //SPI_USR_CMD_HOLD
-#define SPIUADDRHOLD (1 << 21)  //SPI_USR_ADDR_HOLD
-#define SPIUDUMMYHOLD (1 << 20) //SPI_USR_DUMMY_HOLD
-#define SPIUMISOHOLD (1 << 19)  //SPI_USR_DIN_HOLD
-#define SPIUMOSIHOLD (1 << 18)  //SPI_USR_DOUT_HOLD
-#define SPIUHOLDPOL (1 << 17)   //SPI_USR_HOLD_POL
-#define SPIUSIO (1 << 16)		//SPI_SIO
-#define SPIUFWQIO (1 << 15)		//SPI_FWRITE_QIO
-#define SPIUFWDIO (1 << 14)		//SPI_FWRITE_DIO
-#define SPIUFWQUAD (1 << 13)	//SPI_FWRITE_QUAD
-#define SPIUFWDUAL (1 << 12)	//SPI_FWRITE_DUAL
-#define SPIUWRBYO (1 << 11)		//SPI_WR_BYTE_ORDER
-#define SPIURDBYO (1 << 10)		//SPI_RD_BYTE_ORDER
-#define SPIUAHBEM 0x3			//SPI_AHB_ENDIAN_MODE
-#define SPIUAHBEM_S 8			//SPI_AHB_ENDIAN_MODE_S
-#define SPIUSME (1 << 7)		//SPI Master Edge (0:falling, 1:rising), SPI_CK_OUT_EDGE
-#define SPIUSSE (1 << 6)		//SPI Slave Edge (0:falling, 1:rising), SPI_CK_I_EDGE
-#define SPIUCSSETUP (1 << 5)	//SPI_CS_SETUP
-#define SPIUCSHOLD (1 << 4)		//SPI_CS_HOLD
-#define SPIUAHBUCMD (1 << 3)	//SPI_AHB_USR_COMMAND
-#define SPIUAHBUCMD4B (1 << 1)  //SPI_AHB_USR_COMMAND_4BYTE
-#define SPIUDUPLEX (1 << 0)		//SPI_DOUTDIN
+#define SPIUCOMMAND (1 << 31)		//COMMAND pahse, SPI_USR_COMMAND
+#define SPIUADDR (1 << 30)			//ADDRESS phase, SPI_FLASH_USR_ADDR
+#define SPIUDUMMY (1 << 29)			//DUMMY phase, SPI_FLASH_USR_DUMMY
+#define SPIUMISO (1 << 28)			//MISO phase, SPI_FLASH_USR_DIN
+#define SPIUMOSI (1 << 27)			//MOSI phase, SPI_FLASH_DOUT
+#define SPIUDUMMYIDLE (1 << 26)		//SPI_USR_DUMMY_IDLE
+#define SPIUMOSIH (1 << 25)			//MOSI phase uses W8-W15, SPI_USR_DOUT_HIGHPART
+#define SPIUMISOH (1 << 24)			//MISO pahse uses W8-W15, SPI_USR_DIN_HIGHPART
+#define SPIUPREPHOLD (1 << 23)		//SPI_USR_PREP_HOLD
+#define SPIUCMDHOLD (1 << 22)		//SPI_USR_CMD_HOLD
+#define SPIUADDRHOLD (1 << 21)		//SPI_USR_ADDR_HOLD
+#define SPIUDUMMYHOLD (1 << 20)		//SPI_USR_DUMMY_HOLD
+#define SPIUMISOHOLD (1 << 19)		//SPI_USR_DIN_HOLD
+#define SPIUMOSIHOLD (1 << 18)		//SPI_USR_DOUT_HOLD
+#define SPIUHOLDPOL (1 << 17)		//SPI_USR_HOLD_POL
+#define SPIUSIO (1 << 16)			//SPI_SIO
+#define SPIUFWQIO (1 << 15)			//SPI_FWRITE_QIO
+#define SPIUFWDIO (1 << 14)			//SPI_FWRITE_DIO
+#define SPIUFWQUAD (1 << 13)		//SPI_FWRITE_QUAD
+#define SPIUFWDUAL (1 << 12)		//SPI_FWRITE_DUAL
+#define SPIUWRBYO (1 << 11)			//SPI_WR_BYTE_ORDER
+#define SPIURDBYO (1 << 10)			//SPI_RD_BYTE_ORDER
+#define SPIUAHBEM 0x3				//SPI_AHB_ENDIAN_MODE
+#define SPIUAHBEM_S 8				//SPI_AHB_ENDIAN_MODE_S
+#define SPIUSME (1 << 7)			//SPI Master Edge (0:falling, 1:rising), SPI_CK_OUT_EDGE
+#define SPIUSSE (1 << 6)			//SPI Slave Edge (0:falling, 1:rising), SPI_CK_I_EDGE
+#define SPIUCSSETUP (1 << 5)		//SPI_CS_SETUP
+#define SPIUCSHOLD (1 << 4)			//SPI_CS_HOLD
+#define SPIUAHBUCMD (1 << 3)		//SPI_AHB_USR_COMMAND
+#define SPIUAHBUCMD4B (1 << 1)		//SPI_AHB_USR_COMMAND_4BYTE
+#define SPIUDUPLEX (1 << 0)			//SPI_DOUTDIN
 
 //SPI Phase Length Locations
-#define SPILCOMMAND 28 //4 bit in SPIxU2 default 7 (8bit)
-#define SPILADDR 26	//6 bit in SPIxU1 default:23 (24bit)
-#define SPILDUMMY 0	//8 bit in SPIxU1 default:0 (0 cycles)
-#define SPILMISO 8	 //9 bit in SPIxU1 default:0 (1bit)
-#define SPILMOSI 17	//9 bit in SPIxU1 default:0 (1bit)
+#define SPILCOMMAND 28				//4 bit in SPIxU2 default 7 (8bit)
+#define SPILADDR 26					//6 bit in SPIxU1 default:23 (24bit)
+#define SPILDUMMY 0					//8 bit in SPIxU1 default:0 (0 cycles)
+#define SPILMISO 8					//9 bit in SPIxU1 default:0 (1bit)
+#define SPILMOSI 17					//9 bit in SPIxU1 default:0 (1bit)
 //SPI Phase Length Masks
 #define SPIMCOMMAND 0xF
 #define SPIMADDR 0x3F
@@ -505,83 +507,83 @@ extern const uint8_t esp8266_gpioToFn[16];
 #define SPIMMOSI 0x1FF
 
 //SPI Slave (SPIxS)
-#define SPISSRES (1 << 31) //SYNC RESET, SPI_SYNC_RESET
-#define SPISE (1 << 30)	//Slave Enable, SPI_SLAVE_MODE
-#define SPISBE (1 << 29)   //WR/RD BUF enable, SPI_SLV_WR_RD_BUF_EN
-#define SPISSE (1 << 28)   //STA enable, SPI_SLV_WR_RD_STA_EN
-#define SPISCD (1 << 27)   //CMD define, SPI_SLV_CMD_DEFINE
-#define SPISTRCNT 0xF	  //SPI_TRANS_CNT
-#define SPISTRCNT_S 23	 //SPI_TRANS_CNT_S
-#define SPISSLS 0x7		   //SPI_SLV_LAST_STATE
-#define SPISSLS_S 20	   //SPI_SLV_LAST_STATE_S
-#define SPISSLC 0x7		   //SPI_SLV_LAST_COMMAND
-#define SPISSLC_S 17	   //SPI_SLV_LAST_COMMAND_S
-#define SPISCSIM 0x3	   //SPI_CS_I_MODE
-#define SPIDCSIM_S 10	  //SPI_CS_I_MODE_S
-#define SPISTRIE (1 << 9)  //TRANS interrupt enable
-#define SPISWSIE (1 << 8)  //WR_STA interrupt enable
-#define SPISRSIE (1 << 7)  //RD_STA interrupt enable
-#define SPISWBIE (1 << 6)  //WR_BUF interrupt enable
-#define SPISRBIE (1 << 5)  //RD_BUF interrupt enable
-#define SPISTRIS (1 << 4)  //TRANS interrupt status
-#define SPISWSIS (1 << 3)  //WR_STA interrupt status
-#define SPISRSIS (1 << 2)  //RD_STA interrupt status
-#define SPISWBIS (1 << 1)  //WR_BUF interrupt status
-#define SPISRBIS (1 << 0)  //RD_BUF interrupt status
+#define SPISSRES (1 << 31)			 //SYNC RESET, SPI_SYNC_RESET
+#define SPISE (1 << 30)				 //Slave Enable, SPI_SLAVE_MODE
+#define SPISBE (1 << 29)			 //WR/RD BUF enable, SPI_SLV_WR_RD_BUF_EN
+#define SPISSE (1 << 28)			 //STA enable, SPI_SLV_WR_RD_STA_EN
+#define SPISCD (1 << 27)			 //CMD define, SPI_SLV_CMD_DEFINE
+#define SPISTRCNT 0xF				 //SPI_TRANS_CNT
+#define SPISTRCNT_S 23				 //SPI_TRANS_CNT_S
+#define SPISSLS 0x7					 //SPI_SLV_LAST_STATE
+#define SPISSLS_S 20				 //SPI_SLV_LAST_STATE_S
+#define SPISSLC 0x7					 //SPI_SLV_LAST_COMMAND
+#define SPISSLC_S 17				 //SPI_SLV_LAST_COMMAND_S
+#define SPISCSIM 0x3				 //SPI_CS_I_MODE
+#define SPIDCSIM_S 10				 //SPI_CS_I_MODE_S
+#define SPISTRIE (1 << 9)			 //TRANS interrupt enable
+#define SPISWSIE (1 << 8)			 //WR_STA interrupt enable
+#define SPISRSIE (1 << 7)			 //RD_STA interrupt enable
+#define SPISWBIE (1 << 6)			 //WR_BUF interrupt enable
+#define SPISRBIE (1 << 5)			 //RD_BUF interrupt enable
+#define SPISTRIS (1 << 4)			 //TRANS interrupt status
+#define SPISWSIS (1 << 3)			 //WR_STA interrupt status
+#define SPISRSIS (1 << 2)			 //RD_STA interrupt status
+#define SPISWBIS (1 << 1)			 //WR_BUF interrupt status
+#define SPISRBIS (1 << 0)			 //RD_BUF interrupt status
 
 //SPI Slave1 (SPIxS1)
-#define SPIS1LSTA 27		//5 bit in SPIxS1 default:0 (1bit), SPI_SLV_STATUS_BITLEN
-#define SPIS1FE (1 << 26)   //SPI_SLV_STATUS_FAST_EN
-#define SPIS1RSTA (1 << 25) //default:0 enable STA read from Master, SPI_SLV_STATUS_READBACK
-#define SPIS1LBUF 16		//9 bit in SPIxS1 default:0 (1bit), SPI_SLV_BUF_BITLEN
-#define SPIS1LRBA 10		//6 bit in SPIxS1 default:0 (1bit), SPI_SLV_RD_ADDR_BITLEN
-#define SPIS1LWBA 4			//6 bit in SPIxS1 default:0 (1bit), SPI_SLV_WR_ADDR_BITLEN
-#define SPIS1WSDE (1 << 3)  //SPI_SLV_WRSTA_DUMMY_EN
-#define SPIS1RSDE (1 << 2)  //SPI_SLV_RDSTA_DUMMY_EN
-#define SPIS1WBDE (1 << 1)  //SPI_SLV_WRBUF_DUMMY_EN
-#define SPIS1RBDE (1 << 0)  //SPI_SLV_RDBUF_DUMMY_EN
+#define SPIS1LSTA 27				 //5 bit in SPIxS1 default:0 (1bit), SPI_SLV_STATUS_BITLEN
+#define SPIS1FE (1 << 26)			 //SPI_SLV_STATUS_FAST_EN
+#define SPIS1RSTA (1 << 25)			 //default:0 enable STA read from Master, SPI_SLV_STATUS_READBACK
+#define SPIS1LBUF 16				 //9 bit in SPIxS1 default:0 (1bit), SPI_SLV_BUF_BITLEN
+#define SPIS1LRBA 10				 //6 bit in SPIxS1 default:0 (1bit), SPI_SLV_RD_ADDR_BITLEN
+#define SPIS1LWBA 4					 //6 bit in SPIxS1 default:0 (1bit), SPI_SLV_WR_ADDR_BITLEN
+#define SPIS1WSDE (1 << 3)			 //SPI_SLV_WRSTA_DUMMY_EN
+#define SPIS1RSDE (1 << 2)			 //SPI_SLV_RDSTA_DUMMY_EN
+#define SPIS1WBDE (1 << 1)			 //SPI_SLV_WRBUF_DUMMY_EN
+#define SPIS1RBDE (1 << 0)			 //SPI_SLV_RDBUF_DUMMY_EN
 
 //SPI Slave2 (SPIxS2)
-#define SPIS2WBDL 0xFF //SPI_SLV_WRBUF_DUMMY_CYCLELEN
-#define SPIS2WBDL_S 24 //SPI_SLV_WRBUF_DUMMY_CYCLELEN_S
-#define SPIS2RBDL 0xFF //SPI_SLV_RDBUF_DUMMY_CYCLELEN
-#define SPIS2RBDL_S 16 //SPI_SLV_RDBUF_DUMMY_CYCLELEN_S
-#define SPIS2WSDL 0xFF //SPI_SLV_WRSTA_DUMMY_CYCLELEN
-#define SPIS2WSDL_S 8  //SPI_SLV_WRSTA_DUMMY_CYCLELEN_S
-#define SPIS2RSDL 0xFF //SPI_SLV_RDSTA_DUMMY_CYCLELEN
-#define SPIS2RSDL_S 0  //SPI_SLV_RDSTA_DUMMY_CYCLELEN_S
+#define SPIS2WBDL 0xFF				 //SPI_SLV_WRBUF_DUMMY_CYCLELEN
+#define SPIS2WBDL_S 24				 //SPI_SLV_WRBUF_DUMMY_CYCLELEN_S
+#define SPIS2RBDL 0xFF				 //SPI_SLV_RDBUF_DUMMY_CYCLELEN
+#define SPIS2RBDL_S 16				 //SPI_SLV_RDBUF_DUMMY_CYCLELEN_S
+#define SPIS2WSDL 0xFF				 //SPI_SLV_WRSTA_DUMMY_CYCLELEN
+#define SPIS2WSDL_S 8				 //SPI_SLV_WRSTA_DUMMY_CYCLELEN_S
+#define SPIS2RSDL 0xFF				 //SPI_SLV_RDSTA_DUMMY_CYCLELEN
+#define SPIS2RSDL_S 0				 //SPI_SLV_RDSTA_DUMMY_CYCLELEN_S
 
 //SPI Slave3 (SPIxS3)
-#define SPIS3WSCV 0xFF //SPI_SLV_WRSTA_CMD_VALUE
-#define SPIS3WSCV_S 24 //SPI_SLV_WRSTA_CMD_VALUE_S
-#define SPIS3RSCV 0xFF //SPI_SLV_RDSTA_CMD_VALUE
-#define SPIS3RSCV_S 16 //SPI_SLV_RDSTA_CMD_VALUE_S
-#define SPIS3WBCV 0xFF //SPI_SLV_WRBUF_CMD_VALUE
-#define SPIS3WBCV_S 8  //SPI_SLV_WRBUF_CMD_VALUE_S
-#define SPIS3RBCV 0xFF //SPI_SLV_RDBUF_CMD_VALUE
-#define SPIS3RBCV_S 0  //SPI_SLV_RDBUF_CMD_VALUE_S
+#define SPIS3WSCV 0xFF				 //SPI_SLV_WRSTA_CMD_VALUE
+#define SPIS3WSCV_S 24				 //SPI_SLV_WRSTA_CMD_VALUE_S
+#define SPIS3RSCV 0xFF				 //SPI_SLV_RDSTA_CMD_VALUE
+#define SPIS3RSCV_S 16				 //SPI_SLV_RDSTA_CMD_VALUE_S
+#define SPIS3WBCV 0xFF				 //SPI_SLV_WRBUF_CMD_VALUE
+#define SPIS3WBCV_S 8				 //SPI_SLV_WRBUF_CMD_VALUE_S
+#define SPIS3RBCV 0xFF				 //SPI_SLV_RDBUF_CMD_VALUE
+#define SPIS3RBCV_S 0				 //SPI_SLV_RDBUF_CMD_VALUE_S
 
 //SPI EXT0 (SPIxE0)
-#define SPIE0TPPEN (1 << 31) //SPI_T_PP_ENA
-#define SPIE0TPPS 0xF		 //SPI_T_PP_SHIFT
-#define SPIE0TPPS_S 16		 //SPI_T_PP_SHIFT_S
-#define SPIE0TPPT 0xFFF		 //SPI_T_PP_TIME
-#define SPIE0TPPT_S 0		 //SPI_T_PP_TIME_S
+#define SPIE0TPPEN (1 << 31)		 //SPI_T_PP_ENA
+#define SPIE0TPPS 0xF				 //SPI_T_PP_SHIFT
+#define SPIE0TPPS_S 16				 //SPI_T_PP_SHIFT_S
+#define SPIE0TPPT 0xFFF				 //SPI_T_PP_TIME
+#define SPIE0TPPT_S 0				 //SPI_T_PP_TIME_S
 
 //SPI EXT1 (SPIxE1)
-#define SPIE1TEREN (1 << 31) //SPI_T_ERASE_ENA
-#define SPIE1TERS 0xF		 //SPI_T_ERASE_SHIFT
-#define SPIE1TERS_S 16		 //SPI_T_ERASE_SHIFT_S
-#define SPIE1TERT 0xFFF		 //SPI_T_ERASE_TIME
-#define SPIE1TERT_S 0		 //SPI_T_ERASE_TIME_S
+#define SPIE1TEREN (1 << 31)		 //SPI_T_ERASE_ENA
+#define SPIE1TERS 0xF				 //SPI_T_ERASE_SHIFT
+#define SPIE1TERS_S 16				 //SPI_T_ERASE_SHIFT_S
+#define SPIE1TERT 0xFFF				 //SPI_T_ERASE_TIME
+#define SPIE1TERT_S 0				 //SPI_T_ERASE_TIME_S
 
 //SPI EXT2 (SPIxE2)
-#define SPIE2ST 0x7 //SPI_ST
-#define SPIE2ST_S 0 //SPI_ST_S
+#define SPIE2ST 0x7					 //SPI_ST
+#define SPIE2ST_S 0					 //SPI_ST_S
 
 //SPI EXT3 (SPIxE3)
-#define SPIE2IHEN 0x3 //SPI_INT_HOLD_ENA
-#define SPIE2IHEN_S 0 //SPI_INT_HOLD_ENA_S
+#define SPIE2IHEN 0x3				 //SPI_INT_HOLD_ENA
+#define SPIE2IHEN_S 0				 //SPI_INT_HOLD_ENA_S
 
 //SLC (DMA) Registers
 #define SLCC0 ESP8266_REG(0xB00)	 //SLC_CONF0
@@ -628,115 +630,115 @@ extern const uint8_t esp8266_gpioToFn[16];
 #define SLCHC5 ESP8266_REG(0xBBC)	//SLC_HOST_CONF_W5
 
 //SLC (DMA) CONF0
-#define SLCMM (0x3)		  //SLC_MODE
-#define SLCM (12)		  //SLC_MODE_S
-#define SLCDTBE (1 << 9)  //SLC_DATA_BURST_EN
-#define SLCDBE (1 << 8)   //SLC_DSCR_BURST_EN
-#define SLCRXNRC (1 << 7) //SLC_RX_NO_RESTART_CLR
-#define SLCRXAW (1 << 6)  //SLC_RX_AUTO_WRBACK
-#define SLCRXLT (1 << 5)  //SLC_RX_LOOP_TEST
-#define SLCTXLT (1 << 4)  //SLC_TX_LOOP_TEST
-#define SLCAR (1 << 3)	//SLC_AHBM_RST
-#define SLCAFR (1 << 2)   //SLC_AHBM_FIFO_RST
-#define SLCRXLR (1 << 1)  //SLC_RXLINK_RST
-#define SLCTXLR (1 << 0)  //SLC_TXLINK_RST
+#define SLCMM (0x3)					 //SLC_MODE
+#define SLCM (12)					 //SLC_MODE_S
+#define SLCDTBE (1 << 9)			 //SLC_DATA_BURST_EN
+#define SLCDBE (1 << 8)				 //SLC_DSCR_BURST_EN
+#define SLCRXNRC (1 << 7)			 //SLC_RX_NO_RESTART_CLR
+#define SLCRXAW (1 << 6)			 //SLC_RX_AUTO_WRBACK
+#define SLCRXLT (1 << 5)			 //SLC_RX_LOOP_TEST
+#define SLCTXLT (1 << 4)			 //SLC_TX_LOOP_TEST
+#define SLCAR (1 << 3)				 //SLC_AHBM_RST
+#define SLCAFR (1 << 2)				 //SLC_AHBM_FIFO_RST
+#define SLCRXLR (1 << 1)			 //SLC_RXLINK_RST
+#define SLCTXLR (1 << 0)			 //SLC_TXLINK_RST
 
 //SLC (DMA) INT
-#define SLCITXDE (1 << 21)  //SLC_TX_DSCR_EMPTY_INT
-#define SLCIRXDER (1 << 20) //SLC_RX_DSCR_ERR_INT
-#define SLCITXDER (1 << 19) //SLC_TX_DSCR_ERR_INT
-#define SLCITH (1 << 18)	//SLC_TOHOST_INT
-#define SLCIRXEOF (1 << 17) //SLC_RX_EOF_INT
-#define SLCIRXD (1 << 16)   //SLC_RX_DONE_INT
-#define SLCITXEOF (1 << 15) //SLC_TX_EOF_INT
-#define SLCITXD (1 << 14)   //SLC_TX_DONE_INT
-#define SLCIT0 (1 << 13)	//SLC_TOKEN1_1TO0_INT
-#define SLCIT1 (1 << 12)	//SLC_TOKEN0_1TO0_INT
-#define SLCITXO (1 << 11)   //SLC_TX_OVF_INT
-#define SLCIRXU (1 << 10)   //SLC_RX_UDF_INT
-#define SLCITXS (1 << 9)	//SLC_TX_START_INT
-#define SLCIRXS (1 << 8)	//SLC_RX_START_INT
-#define SLCIFH7 (1 << 7)	//SLC_FRHOST_BIT7_INT
-#define SLCIFH6 (1 << 6)	//SLC_FRHOST_BIT6_INT
-#define SLCIFH5 (1 << 5)	//SLC_FRHOST_BIT5_INT
-#define SLCIFH4 (1 << 4)	//SLC_FRHOST_BIT4_INT
-#define SLCIFH3 (1 << 3)	//SLC_FRHOST_BIT3_INT
-#define SLCIFH2 (1 << 2)	//SLC_FRHOST_BIT2_INT
-#define SLCIFH1 (1 << 1)	//SLC_FRHOST_BIT1_INT
-#define SLCIFH0 (1 << 0)	//SLC_FRHOST_BIT0_INT
+#define SLCITXDE (1 << 21)			 //SLC_TX_DSCR_EMPTY_INT
+#define SLCIRXDER (1 << 20)			 //SLC_RX_DSCR_ERR_INT
+#define SLCITXDER (1 << 19)			 //SLC_TX_DSCR_ERR_INT
+#define SLCITH (1 << 18)			 //SLC_TOHOST_INT
+#define SLCIRXEOF (1 << 17)			 //SLC_RX_EOF_INT
+#define SLCIRXD (1 << 16)			 //SLC_RX_DONE_INT
+#define SLCITXEOF (1 << 15)			 //SLC_TX_EOF_INT
+#define SLCITXD (1 << 14)			 //SLC_TX_DONE_INT
+#define SLCIT0 (1 << 13)			 //SLC_TOKEN1_1TO0_INT
+#define SLCIT1 (1 << 12)			 //SLC_TOKEN0_1TO0_INT
+#define SLCITXO (1 << 11)			 //SLC_TX_OVF_INT
+#define SLCIRXU (1 << 10)			 //SLC_RX_UDF_INT
+#define SLCITXS (1 << 9)			 //SLC_TX_START_INT
+#define SLCIRXS (1 << 8)			 //SLC_RX_START_INT
+#define SLCIFH7 (1 << 7)			 //SLC_FRHOST_BIT7_INT
+#define SLCIFH6 (1 << 6)			 //SLC_FRHOST_BIT6_INT
+#define SLCIFH5 (1 << 5)			 //SLC_FRHOST_BIT5_INT
+#define SLCIFH4 (1 << 4)			 //SLC_FRHOST_BIT4_INT
+#define SLCIFH3 (1 << 3)			 //SLC_FRHOST_BIT3_INT
+#define SLCIFH2 (1 << 2)			 //SLC_FRHOST_BIT2_INT
+#define SLCIFH1 (1 << 1)			 //SLC_FRHOST_BIT1_INT
+#define SLCIFH0 (1 << 0)			 //SLC_FRHOST_BIT0_INT
 
 //SLC (DMA) RX_STATUS
-#define SLCRXE (1 << 1) //SLC_RX_EMPTY
-#define SLCRXF (1 << 0) //SLC_RX_FULL
+#define SLCRXE (1 << 1)				 //SLC_RX_EMPTY
+#define SLCRXF (1 << 0)				 //SLC_RX_FULL
 
 //SLC (DMA) TX_STATUS
-#define SLCTXE (1 << 1) //SLC_TX_EMPTY
-#define SLCTXF (1 << 0) //SLC_TX_FULL
+#define SLCTXE (1 << 1)				 //SLC_TX_EMPTY
+#define SLCTXF (1 << 0)				 //SLC_TX_FULL
 
 //SLC (DMA) RX_FIFO_PUSH
-#define SLCRXFP (1 << 16) //SLC_RXFIFO_PUSH
-#define SLCRXWDM (0x1FF)  //SLC_RXFIFO_WDATA
-#define SLCRXWD (0)		  //SLC_RXFIFO_WDATA_S
+#define SLCRXFP (1 << 16)			 //SLC_RXFIFO_PUSH
+#define SLCRXWDM (0x1FF)			 //SLC_RXFIFO_WDATA
+#define SLCRXWD (0)					 //SLC_RXFIFO_WDATA_S
 
 //SLC (DMA) TX_FIFO_POP
-#define SLCTXFP (1 << 16) //SLC_TXFIFO_POP
-#define SLCTXRDM (0x7FF)  //SLC_TXFIFO_RDATA
-#define SLCTXRD (0)		  //SLC_TXFIFO_RDATA_S
+#define SLCTXFP (1 << 16)			 //SLC_TXFIFO_POP
+#define SLCTXRDM (0x7FF)			 //SLC_TXFIFO_RDATA
+#define SLCTXRD (0)					 //SLC_TXFIFO_RDATA_S
 
 //SLC (DMA) RX_LINK
-#define SLCRXLP (1 << 31)  //SLC_RXLINK_PARK
-#define SLCRXLRS (1 << 30) //SLC_RXLINK_RESTART
-#define SLCRXLS (1 << 29)  //SLC_RXLINK_START
-#define SLCRXLE (1 << 28)  //SLC_RXLINK_STOP
-#define SLCRXLAM (0xFFFF)  //SLC_RXLINK_DESCADDR_MASK
-#define SLCRXLA (0)		   //SLC_RXLINK_ADDR_S
+#define SLCRXLP (1 << 31)			 //SLC_RXLINK_PARK
+#define SLCRXLRS (1 << 30)			 //SLC_RXLINK_RESTART
+#define SLCRXLS (1 << 29)			 //SLC_RXLINK_START
+#define SLCRXLE (1 << 28)			 //SLC_RXLINK_STOP
+#define SLCRXLAM (0xFFFF)			 //SLC_RXLINK_DESCADDR_MASK
+#define SLCRXLA (0)					 //SLC_RXLINK_ADDR_S
 
 //SLC (DMA) TX_LINK
-#define SLCTXLP (1 << 31)  //SLC_TXLINK_PARK
-#define SLCTXLRS (1 << 30) //SLC_TXLINK_RESTART
-#define SLCTXLS (1 << 29)  //SLC_TXLINK_START
-#define SLCTXLE (1 << 28)  //SLC_TXLINK_STOP
-#define SLCTXLAM (0xFFFF)  //SLC_TXLINK_DESCADDR_MASK
-#define SLCTXLA (0)		   //SLC_TXLINK_ADDR_S
+#define SLCTXLP (1 << 31)			 //SLC_TXLINK_PARK
+#define SLCTXLRS (1 << 30)			 //SLC_TXLINK_RESTART
+#define SLCTXLS (1 << 29)			 //SLC_TXLINK_START
+#define SLCTXLE (1 << 28)			 //SLC_TXLINK_STOP
+#define SLCTXLAM (0xFFFF)			 //SLC_TXLINK_DESCADDR_MASK
+#define SLCTXLA (0)					 //SLC_TXLINK_ADDR_S
 
 //SLC (DMA) TOKENx
-#define SLCTM (0xFFF)	//SLC_TOKENx_MASK
-#define SLCTT (16)		 //SLC_TOKENx_S
-#define SLCTIM (1 << 14) //SLC_TOKENx_LOCAL_INC_MORE
-#define SLCTI (1 << 13)  //SLC_TOKENx_LOCAL_INC
-#define SLCTW (1 << 12)  //SLC_TOKENx_LOCAL_WR
-#define SLCTDM (0xFFF)   //SLC_TOKENx_LOCAL_WDATA
-#define SLCTD (0)		 //SLC_TOKENx_LOCAL_WDATA_S
+#define SLCTM (0xFFF)				 //SLC_TOKENx_MASK
+#define SLCTT (16)					 //SLC_TOKENx_S
+#define SLCTIM (1 << 14)			 //SLC_TOKENx_LOCAL_INC_MORE
+#define SLCTI (1 << 13)				 //SLC_TOKENx_LOCAL_INC
+#define SLCTW (1 << 12)				 //SLC_TOKENx_LOCAL_WR
+#define SLCTDM (0xFFF)				 //SLC_TOKENx_LOCAL_WDATA
+#define SLCTD (0)					 //SLC_TOKENx_LOCAL_WDATA_S
 
 //SLC (DMA) BRIDGE_CONF
-#define SLCBFMEM (0xF)  //SLC_FIFO_MAP_ENA
-#define SLCBFME (8)		//SLC_FIFO_MAP_ENA_S
-#define SLCBTEEM (0x3F) //SLC_TXEOF_ENA
-#define SLCBTEE (0)		//SLC_TXEOF_ENA_S
+#define SLCBFMEM (0xF)				 //SLC_FIFO_MAP_ENA
+#define SLCBFME (8)					 //SLC_FIFO_MAP_ENA_S
+#define SLCBTEEM (0x3F)				 //SLC_TXEOF_ENA
+#define SLCBTEE (0)					 //SLC_TXEOF_ENA_S
 
 //SLC (DMA) AHB_TEST
-#define SLCATAM (0x3) //SLC_AHB_TESTADDR
-#define SLCATA (4)	//SLC_AHB_TESTADDR_S
-#define SLCATMM (0x7) //SLC_AHB_TESTMODE
-#define SLCATM (0)	//SLC_AHB_TESTMODE_S
+#define SLCATAM (0x3)				 //SLC_AHB_TESTADDR
+#define SLCATA (4)					 //SLC_AHB_TESTADDR_S
+#define SLCATMM (0x7)				 //SLC_AHB_TESTMODE
+#define SLCATM (0)					 //SLC_AHB_TESTMODE_S
 
 //SLC (DMA) SDIO_ST
-#define SLCSBM (0x7)   //SLC_BUS_ST
-#define SLCSB (12)	 //SLC_BUS_ST_S
-#define SLCSW (1 << 8) //SLC_SDIO_WAKEUP
-#define SLCSFM (0xF)   //SLC_FUNC_ST
-#define SLCSF (4)	  //SLC_FUNC_ST_S
-#define SLCSCM (0x7)   //SLC_CMD_ST
-#define SLCSC (0)	  //SLC_CMD_ST_S
+#define SLCSBM (0x7)				 //SLC_BUS_ST
+#define SLCSB (12)					 //SLC_BUS_ST_S
+#define SLCSW (1 << 8)				 //SLC_SDIO_WAKEUP
+#define SLCSFM (0xF)				 //SLC_FUNC_ST
+#define SLCSF (4)					 //SLC_FUNC_ST_S
+#define SLCSCM (0x7)				 //SLC_CMD_ST
+#define SLCSC (0)					 //SLC_CMD_ST_S
 
 //SLC (DMA) RX_DSCR_CONF
-#define SLCBRXFE (1 << 20) //SLC_RX_FILL_EN
-#define SLCBRXEM (1 << 19) //SLC_RX_EOF_MODE
-#define SLCBRXFM (1 << 18) //SLC_RX_FILL_MODE
-#define SLCBINR (1 << 17)  //SLC_INFOR_NO_REPLACE
-#define SLCBTNR (1 << 16)  //SLC_TOKEN_NO_REPLACE
-#define SLCBPICM (0xFFFF)  //SLC_POP_IDLE_CNT
-#define SLCBPIC (0)		   //SLC_POP_IDLE_CNT_S
+#define SLCBRXFE (1 << 20)			 //SLC_RX_FILL_EN
+#define SLCBRXEM (1 << 19)			 //SLC_RX_EOF_MODE
+#define SLCBRXFM (1 << 18)			 //SLC_RX_FILL_MODE
+#define SLCBINR (1 << 17)			 //SLC_INFOR_NO_REPLACE
+#define SLCBTNR (1 << 16)			 //SLC_TOKEN_NO_REPLACE
+#define SLCBPICM (0xFFFF)			 //SLC_POP_IDLE_CNT
+#define SLCBPIC (0)					 //SLC_POP_IDLE_CNT_S
 
 // I2S Registers
 #define i2c_bbpll 0x67
@@ -761,75 +763,75 @@ extern const uint8_t esp8266_gpioToFn[16];
 #define I2SCC ESP8266_REG(0xe2C)   //I2SCONF_CHAN
 
 // I2S CONF
-#define I2SBDM (0x3F)	//I2S_BCK_DIV_NUM
-#define I2SBD (22)		 //I2S_BCK_DIV_NUM_S
-#define I2SCDM (0x3F)	//I2S_CLKM_DIV_NUM
-#define I2SCD (16)		 //I2S_CLKM_DIV_NUM_S
-#define I2SBMM (0xF)	 //I2S_BITS_MOD
-#define I2SBM (12)		 //I2S_BITS_MOD_S
-#define I2SRMS (1 << 11) //I2S_RECE_MSB_SHIFT
-#define I2STMS (1 << 10) //I2S_TRANS_MSB_SHIFT
-#define I2SRXS (1 << 9)  //I2S_I2S_RX_START
-#define I2STXS (1 << 8)  //I2S_I2S_TX_START
-#define I2SMR (1 << 7)   //I2S_MSB_RIGHT
-#define I2SRF (1 << 6)   //I2S_RIGHT_FIRST
-#define I2SRSM (1 << 5)  //I2S_RECE_SLAVE_MOD
-#define I2STSM (1 << 4)  //I2S_TRANS_SLAVE_MOD
-#define I2SRXFR (1 << 3) //I2S_I2S_RX_FIFO_RESET
-#define I2STXFR (1 << 2) //I2S_I2S_TX_FIFO_RESET
-#define I2SRXR (1 << 1)  //I2S_I2S_RX_RESET
-#define I2STXR (1 << 0)  //I2S_I2S_TX_RESET
-#define I2SRST (0xF)	 //I2S_I2S_RESET_MASK
+#define I2SBDM (0x3F)			   //I2S_BCK_DIV_NUM
+#define I2SBD (22)				   //I2S_BCK_DIV_NUM_S
+#define I2SCDM (0x3F)			   //I2S_CLKM_DIV_NUM
+#define I2SCD (16)				   //I2S_CLKM_DIV_NUM_S
+#define I2SBMM (0xF)			   //I2S_BITS_MOD
+#define I2SBM (12)				   //I2S_BITS_MOD_S
+#define I2SRMS (1 << 11)		   //I2S_RECE_MSB_SHIFT
+#define I2STMS (1 << 10)		   //I2S_TRANS_MSB_SHIFT
+#define I2SRXS (1 << 9)			   //I2S_I2S_RX_START
+#define I2STXS (1 << 8)			   //I2S_I2S_TX_START
+#define I2SMR (1 << 7)			   //I2S_MSB_RIGHT
+#define I2SRF (1 << 6)			   //I2S_RIGHT_FIRST
+#define I2SRSM (1 << 5)			   //I2S_RECE_SLAVE_MOD
+#define I2STSM (1 << 4)			   //I2S_TRANS_SLAVE_MOD
+#define I2SRXFR (1 << 3)		   //I2S_I2S_RX_FIFO_RESET
+#define I2STXFR (1 << 2)		   //I2S_I2S_TX_FIFO_RESET
+#define I2SRXR (1 << 1)			   //I2S_I2S_RX_RESET
+#define I2STXR (1 << 0)			   //I2S_I2S_TX_RESET
+#define I2SRST (0xF)			   //I2S_I2S_RESET_MASK
 
 //I2S INT
-#define I2SITXRE (1 << 5) //I2S_I2S_TX_REMPTY_INT
-#define I2SITXWF (1 << 4) //I2S_I2S_TX_WFULL_INT
-#define I2SIRXRE (1 << 3) //I2S_I2S_RX_REMPTY_INT
-#define I2SIRXWF (1 << 2) //I2S_I2S_RX_WFULL_INT
-#define I2SITXPD (1 << 1) //I2S_I2S_TX_PUT_DATA_INT
-#define I2SIRXTD (1 << 0) //I2S_I2S_RX_TAKE_DATA_INT
+#define I2SITXRE (1 << 5)		   //I2S_I2S_TX_REMPTY_INT
+#define I2SITXWF (1 << 4)		   //I2S_I2S_TX_WFULL_INT
+#define I2SIRXRE (1 << 3)		   //I2S_I2S_RX_REMPTY_INT
+#define I2SIRXWF (1 << 2)		   //I2S_I2S_RX_WFULL_INT
+#define I2SITXPD (1 << 1)		   //I2S_I2S_TX_PUT_DATA_INT
+#define I2SIRXTD (1 << 0)		   //I2S_I2S_RX_TAKE_DATA_INT
 
 //I2S TIMING
-#define I2STBII (1 << 22) //I2S_TRANS_BCK_IN_INV
-#define I2SRDS (1 << 21)  //I2S_RECE_DSYNC_SW
-#define I2STDS (1 << 20)  //I2S_TRANS_DSYNC_SW
-#define I2SRBODM (0x3)	//I2S_RECE_BCK_OUT_DELAY
-#define I2SRBOD (18)	  //I2S_RECE_BCK_OUT_DELAY_S
-#define I2SRWODM (0x3)	//I2S_RECE_WS_OUT_DELAY
-#define I2SRWOD (16)	  //I2S_RECE_WS_OUT_DELAY_S
-#define I2STSODM (0x3)	//I2S_TRANS_SD_OUT_DELAY
-#define I2STSOD (14)	  //I2S_TRANS_SD_OUT_DELAY_S
-#define I2STWODM (0x3)	//I2S_TRANS_WS_OUT_DELAY
-#define I2STWOD (12)	  //I2S_TRANS_WS_OUT_DELAY_S
-#define I2STBODM (0x3)	//I2S_TRANS_BCK_OUT_DELAY
-#define I2STBOD (10)	  //I2S_TRANS_BCK_OUT_DELAY_S
-#define I2SRSIDM (0x3)	//I2S_RECE_SD_IN_DELAY
-#define I2SRSID (8)		  //I2S_RECE_SD_IN_DELAY_S
-#define I2SRWIDM (0x3)	//I2S_RECE_WS_IN_DELAY
-#define I2SRWID (6)		  //I2S_RECE_WS_IN_DELAY_S
-#define I2SRBIDM (0x3)	//I2S_RECE_BCK_IN_DELAY
-#define I2SRBID (4)		  //I2S_RECE_BCK_IN_DELAY_S
-#define I2STWIDM (0x3)	//I2S_TRANS_WS_IN_DELAY
-#define I2STWID (2)		  //I2S_TRANS_WS_IN_DELAY_S
-#define I2STBIDM (0x3)	//I2S_TRANS_BCK_IN_DELAY
-#define I2STBID (0)		  //I2S_TRANS_BCK_IN_DELAY_S
+#define I2STBII (1 << 22)		   //I2S_TRANS_BCK_IN_INV
+#define I2SRDS (1 << 21)		   //I2S_RECE_DSYNC_SW
+#define I2STDS (1 << 20)		   //I2S_TRANS_DSYNC_SW
+#define I2SRBODM (0x3)			   //I2S_RECE_BCK_OUT_DELAY
+#define I2SRBOD (18)			   //I2S_RECE_BCK_OUT_DELAY_S
+#define I2SRWODM (0x3)			   //I2S_RECE_WS_OUT_DELAY
+#define I2SRWOD (16)			   //I2S_RECE_WS_OUT_DELAY_S
+#define I2STSODM (0x3)			   //I2S_TRANS_SD_OUT_DELAY
+#define I2STSOD (14)			   //I2S_TRANS_SD_OUT_DELAY_S
+#define I2STWODM (0x3)			   //I2S_TRANS_WS_OUT_DELAY
+#define I2STWOD (12)			   //I2S_TRANS_WS_OUT_DELAY_S
+#define I2STBODM (0x3)			   //I2S_TRANS_BCK_OUT_DELAY
+#define I2STBOD (10)			   //I2S_TRANS_BCK_OUT_DELAY_S
+#define I2SRSIDM (0x3)			   //I2S_RECE_SD_IN_DELAY
+#define I2SRSID (8)				   //I2S_RECE_SD_IN_DELAY_S
+#define I2SRWIDM (0x3)			   //I2S_RECE_WS_IN_DELAY
+#define I2SRWID (6)				   //I2S_RECE_WS_IN_DELAY_S
+#define I2SRBIDM (0x3)			   //I2S_RECE_BCK_IN_DELAY
+#define I2SRBID (4)				   //I2S_RECE_BCK_IN_DELAY_S
+#define I2STWIDM (0x3)			   //I2S_TRANS_WS_IN_DELAY
+#define I2STWID (2)				   //I2S_TRANS_WS_IN_DELAY_S
+#define I2STBIDM (0x3)			   //I2S_TRANS_BCK_IN_DELAY
+#define I2STBID (0)				   //I2S_TRANS_BCK_IN_DELAY_S
 
 //I2S FIFO CONF
-#define I2SRXFMM (0x7)  //I2S_I2S_RX_FIFO_MOD
-#define I2SRXFM (16)	//I2S_I2S_RX_FIFO_MOD_S
-#define I2STXFMM (0x7)  //I2S_I2S_TX_FIFO_MOD
-#define I2STXFM (13)	//I2S_I2S_TX_FIFO_MOD_S
-#define I2SDE (1 << 12) //I2S_I2S_DSCR_EN
-#define I2STXDNM (0x3F) //I2S_I2S_TX_DATA_NUM
-#define I2STXDN (6)		//I2S_I2S_TX_DATA_NUM_S
-#define I2SRXDNM (0x3F) //I2S_I2S_RX_DATA_NUM
-#define I2SRXDN (0)		//I2S_I2S_RX_DATA_NUM_S
+#define I2SRXFMM (0x7)			   //I2S_I2S_RX_FIFO_MOD
+#define I2SRXFM (16)			   //I2S_I2S_RX_FIFO_MOD_S
+#define I2STXFMM (0x7)			   //I2S_I2S_TX_FIFO_MOD
+#define I2STXFM (13)			   //I2S_I2S_TX_FIFO_MOD_S
+#define I2SDE (1 << 12)			   //I2S_I2S_DSCR_EN
+#define I2STXDNM (0x3F)			   //I2S_I2S_TX_DATA_NUM
+#define I2STXDN (6)				   //I2S_I2S_TX_DATA_NUM_S
+#define I2SRXDNM (0x3F)			   //I2S_I2S_RX_DATA_NUM
+#define I2SRXDN (0)				   //I2S_I2S_RX_DATA_NUM_S
 
 //I2S CONF CHAN
-#define I2SRXCMM (0x3) //I2S_RX_CHAN_MOD
-#define I2SRXCM (3)	//I2S_RX_CHAN_MOD_S
-#define I2STXCMM (0x7) //I2S_TX_CHAN_MOD
-#define I2STXCM (0)	//I2S_TX_CHAN_MOD_S
+#define I2SRXCMM (0x3)			   //I2S_RX_CHAN_MOD
+#define I2SRXCM (3)				   //I2S_RX_CHAN_MOD_S
+#define I2STXCMM (0x7)			   //I2S_TX_CHAN_MOD
+#define I2STXCM (0)				   //I2S_TX_CHAN_MOD_S
 
 /**
  Random Number Generator 32bit
@@ -839,4 +841,6 @@ extern const uint8_t esp8266_gpioToFn[16];
 
 #ifdef __cplusplus
 }
+#endif
+
 #endif
