+++ Sming/Arch/Esp8266/Components/driver/include/driver/os_timer.h
@@ -35,7 +35,7 @@ extern "C" {
  * This function has been added to Sming for more efficient and flexible use of
  * software timers. It can be used alongside the SDK `os_timer_arm_new()` function.
  */
-void IRAM_ATTR os_timer_arm_ticks(os_timer_t* ptimer, uint32_t ticks, bool repeat_flag);
+void os_timer_arm_ticks(os_timer_t* ptimer, uint32_t ticks, bool repeat_flag);
 
 /** @} */
