+++ Sming/Arch/Host/Components/driver/include/driver/os_timer.h
@@ -7,7 +7,8 @@
  */
 
 #pragma once
-#include "c_types.h"
+
+#include <c_types.h>
 
 #ifdef __cplusplus
 extern "C" {
