+++ Sming/Components/arch_driver/README.rst
@@ -1,0 +1,12 @@
+Arch Driver
+===========
+
+This is an internal Component to pull together common headers and code used at a low-level by
+more than once Arch.
+
+This is to:
+
+-  Ease maintenance
+-  Avoid code duplication
+-  Provide a consistent API for the framework to use
+
