+++ Sming/Components/arch_driver/component.mk
@@ -1,0 +1,2 @@
+COMPONENT_INCDIRS := src/include
+COMPONENT_SRCDIRS := src
