+++ tests/HostTests/app/test-serial.cpp
@@ -1,8 +1,6 @@
 #include <SmingTest.h>
 
-#if defined(ARCH_HOST) || defined(ARCH_ESP8266)
-#include <SerialBuffer.h>
-#endif
+#include <driver/SerialBuffer.h>
 
 class SerialTest : public TestGroup
 {
@@ -15,7 +13,6 @@ public:
 	{
 		TEST_CASE("SerialBuffer write/read")
 		{
-#if defined(ARCH_HOST) || defined(ARCH_ESP8266)
 			static constexpr size_t BUFSIZE = 128;
 			SerialBuffer txbuf;
 			txbuf.resize(BUFSIZE);
@@ -58,7 +55,6 @@ public:
 			readBuffer += read();
 			REQUIRE(txbuf.available() == 0);
 			REQUIRE(compareBuffer == readBuffer);
-#endif
 		}
 	}
 };
